Generate Tile method moved to auxiliar class
