+++ DungeonProject/DungeonProject/DungeonProject.py
@@ -68,45 +68,6 @@ def makeBox(width, height, depth, manager, nodeName = "", meshName = ""):
 
     return newNode
 
-# Places a the tile specified by index to match with the point specified by transform, and adds it to the scene.
-# Returns the new transform points where to place the next tiles
-# If the tile to be generated would overlap any of the previously placed tiles, returns false.
-def generateTile(index, transform, manager, kitScene, scene, placedTiles):
-    # Calculates the x and z coordinates of the tile (these depend on the rotation of the "in point" of the tile)
-    x = TileFile.tiles[index][1][0] * MathModule.cos(TileFile.tiles[index][1][3]) + TileFile.tiles[index][1][2] * MathModule.sin(TileFile.tiles[index][1][3])
-    z = TileFile.tiles[index][1][2] * MathModule.cos(TileFile.tiles[index][1][3]) + TileFile.tiles[index][1][0] * MathModule.sin(TileFile.tiles[index][1][3])
-
-    # Sets the final coordinates of the tile to match the coordinates of the given transform
-    newTransform = [transform[0] + x * MathModule.cos(transform[3]) - z * MathModule.sin(transform[3]),
-                    transform[1] - TileFile.tiles[index][1][1],
-                    transform[2] - z * MathModule.cos(transform[3]) + x * MathModule.sin(transform[3]),
-                    transform[3] - TileFile.tiles[index][1][3], 
-                    TileFile.tiles[index][1][4]]
-
-    # Check if the tile would overlap any of the previously place
-    bb = [newTransform, TileFile.tiles[index][3]]
-    if checkCollision(bb, placedTiles):
-        return False
-
-    # Add the tile to the collision system
-    placedTiles += [bb]
-
-    # Create the node with the tile's mesh and adds it to the scene
-    tile = TileGenerator.copyTile(TileFile.tiles[index][0], manager, kitScene)
-    scene2.GetRootNode().AddChild(tile)
-
-    # Set the transform of the tile
-    tile.LclRotation.Set(FbxDouble3(tile.LclRotation.Get()[0], tile.LclRotation.Get()[1] + newTransform[3], tile.LclRotation.Get()[2]))
-    tile.LclTranslation.Set(FbxDouble3(newTransform[0], newTransform[1], newTransform[2]))
-
-    # Returns a list of the new points (transforms) where the next tiles will be placed
-    return [[transform[0] + exit[0] * MathModule.cos(transform[3]) + exit[2] * MathModule.sin(transform[3]),
-            transform[1] + exit[1],
-            transform[2] + exit[2] * MathModule.cos(transform[3]) - exit[0] * MathModule.sin(transform[3]),
-            transform[3] + exit[3],
-            exit[4]]
-            for exit in TileFile.tiles[index][2]]
-
 # Generates a tree graph that represents the high level structure of the dungeon
 # Nodes are represented by each letter character
 # Two nodes connected by one path are represented by two consecutive letters: XX
@@ -307,7 +268,7 @@ def buildRoom(properties, transform, manager, kitScene, scene, placedTiles):
     originalTransform = [i for i in transform]
 
     #Build the room with one tile according to the number of exits needed for the room
-    transform = generateTile({
+    transform = TileGenerator.generateTile({
             0: 12 if originalTransform[4] == 400 or properties["isSpawnRoom"] else 30,
             1: 12 if originalTransform[4] == 400 or properties["isSpawnRoom"] else 30,
             2: 15 if originalTransform[4] == 400 or properties["isSpawnRoom"] else 33,
@@ -319,7 +280,7 @@ def buildRoom(properties, transform, manager, kitScene, scene, placedTiles):
         return False
 
     #Build entry door
-    generateTile({
+    TileGenerator.generateTile({
             400: 22 if not properties["isSpawnRoom"] else 19,
             800: 40 if not properties["isSpawnRoom"] else 37,
             1600: 42 if not properties["isSpawnRoom"] else 39
@@ -327,7 +288,7 @@ def buildRoom(properties, transform, manager, kitScene, scene, placedTiles):
 
     #Build doors on each exit
     for i in range(len(transform)):
-        transform[i] = generateTile({
+        transform[i] = TileGenerator.generateTile({
                 400: 8 if properties["numExits"] > 0 else 9,
                 800: 23 if properties["numExits"] > 0 else 26,
                 1600: 41 if properties["numExits"] > 0 else 44
@@ -388,7 +349,7 @@ def buildPath(transform, manager, kitScene, scene, placedTiles, difficultyLevel)
         # Try with all alternatives
         while not ret and len(weights[-1]) > 0:
             tile = randomWeightedChoice(weights[-1])
-            ret = generateTile(tile, transforms[-1], manager, kitScene, scene, placedTiles)
+            ret = TileGenerator.generateTile(tile, transforms[-1], manager, kitScene, scene, placedTiles)
             del weights[-1][tile]
 
         if ret:
